extracted Rubberduck.Core.AutoComplete namespace
