+++ Rubberduck.Core/AutoComplete/AutoCompleteClosingBrace.cs
@@ -1,0 +1,8 @@
+﻿namespace Rubberduck.AutoComplete
+{
+    public class AutoCompleteClosingBrace : AutoCompleteBase
+    {
+        public override string InputToken => "{";
+        public override string OutputToken => "}";
+    }
+}
