+++ Rubberduck.Core/AutoComplete/AutoCompleteClosingBracket.cs
@@ -1,0 +1,8 @@
+﻿namespace Rubberduck.AutoComplete
+{
+    public class AutoCompleteClosingBracket : AutoCompleteBase
+    {
+        public override string InputToken => "[";
+        public override string OutputToken => "]";
+    }
+}
