+++ Rubberduck.Core/AutoComplete/AutoCompleteClosingParenthese.cs
@@ -1,0 +1,8 @@
+﻿namespace Rubberduck.AutoComplete
+{
+    public class AutoCompleteClosingParenthese : AutoCompleteBase
+    {
+        public override string InputToken => "(";
+        public override string OutputToken => ")";
+    }
+}
