+++ Rubberduck.Core/AutoComplete/AutoCompleteClosingString.cs
@@ -1,0 +1,8 @@
+﻿namespace Rubberduck.AutoComplete
+{
+    public class AutoCompleteClosingString : AutoCompleteBase
+    {
+        public override string InputToken => "\"";
+        public override string OutputToken => "\"";
+    }
+}
