+++ Rubberduck.Core/AutoComplete/AutoCompleteService.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Rubberduck.VBEditor;
+using Rubberduck.VBEditor.Events;
+
+namespace Rubberduck.AutoComplete
+{
+    public class AutoCompleteService : IAutoCompleteService, IDisposable
+    {
+        public event EventHandler TypingCode;
+        private readonly IReadOnlyList<IAutoComplete> _autocompletions = new IAutoComplete[]
+        {
+            new AutoCompleteClosingParenthese(),
+            new AutoCompleteClosingString(),
+            new AutoCompleteClosingBracket(),
+            new AutoCompleteClosingBrace(),
+            new AutoCompleteEndIf(),
+        };
+
+        public AutoCompleteService()
+        {
+            VBENativeServices.TypingCode += VBENativeServices_TypingCode;
+        }
+
+        QualifiedSelection? _lastSelection;
+        string _lastCode;
+
+        private void VBENativeServices_TypingCode(object sender, AutoCompleteEventArgs e)
+        {
+            TypingCode?.Invoke(this, e);
+            var selection = e.CodePane.Selection;
+            var qualifiedSelection = e.CodePane.GetQualifiedSelection();
+
+            if (!selection.IsSingleCharacter || e.Code.Equals(_lastCode) || qualifiedSelection.Value.Equals(_lastSelection))
+            {
+                return;
+            }
+
+            foreach (var autocomplete in _autocompletions.Where(auto => auto.IsEnabled))
+            {
+                if (autocomplete.Execute(e))
+                {
+                    _lastSelection = qualifiedSelection;
+                    _lastCode = e.ReplacementLineContent;
+                    break;
+                }
+            }
+        }
+
+        public void Dispose()
+        {
+            VBENativeServices.TypingCode -= VBENativeServices_TypingCode;
+        }
+    }
+}
