+++ Rubberduck.Core/Rubberduck.Core.csproj
@@ -372,6 +372,13 @@
     <Compile Include="UI\About\AboutDialog.Designer.cs">
       <DependentUpon>AboutDialog.cs</DependentUpon>
     </Compile>
+    <Compile Include="AutoComplete\AutoCompleteBase.cs" />
+    <Compile Include="AutoComplete\AutoCompleteClosingBrace.cs" />
+    <Compile Include="AutoComplete\AutoCompleteClosingBracket.cs" />
+    <Compile Include="AutoComplete\AutoCompleteClosingParenthese.cs" />
+    <Compile Include="AutoComplete\AutoCompleteClosingString.cs" />
+    <Compile Include="AutoComplete\AutoCompleteEndIf.cs" />
+    <Compile Include="AutoComplete\AutoCompleteService.cs" />
     <Compile Include="UI\CodeExplorer\Commands\AddTestModuleWithStubsCommand.cs" />
     <Compile Include="UI\CodeExplorer\Commands\CodeExplorerCommandAttribute.cs" />
     <Compile Include="UI\CodeExplorer\Commands\AddUserFormCommand.cs" />
@@ -420,6 +427,8 @@
     <Compile Include="UI\Converters\SubtractionConverter.cs" />
     <Compile Include="UI\CustomComWrappers\IOleWrapper.cs" />
     <Compile Include="UI\EnvironmentProvider.cs" />
+    <Compile Include="AutoComplete\IAutoComplete.cs" />
+    <Compile Include="UI\IAutoCompleteService.cs" />
     <Compile Include="UI\Inspections\AggregateInspectionResult.cs" />
     <Compile Include="UI\ModernFolderBrowser.cs" />
     <Compile Include="UI\Refactorings\EncapsulateField\EncapsulateFieldView.xaml.cs">
