+++ Rubberduck.Core/UI/IAutoCompleteService.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Rubberduck.AutoComplete
+{
+    public interface IAutoCompleteService
+    {
+        event EventHandler TypingCode;
+    }
+}
